[refactor] rename script util lib
